splitting editor.js into components
