+++ static/js/src/admin/components/Image.jsx
@@ -1,0 +1,16 @@
+var Image = React.createClass({
+    onClick: function(){
+        this.props.onClickHandler(this.props.id);
+    },
+    render: function(){
+        var styles = {backgroundImage: "url('" + this.props.thumb + "')"};
+        if(this.props.progress){
+        //    styles.opacity = 100 / this.props.progress;
+        }
+        return (
+            <li className={'catalog-image'} onClick={this.onClick} style={styles} data-id={this.props.id} data-url={this.props.url}></li>
+        );
+    }
+});
+
+module.exports = Image;
